Add initial motherboard eagle files
Initial Eagle Schematic, Board, and Library files for a Pumpkin CubeSat
Kit-compatible Slot 0 Motherboard.  This is a template, mechanically it
should work, but electrically it doesn't really do much yet.  This is the
board outline and connectors for the main modules: a 104-pin CSK bus
with H1 and H2 connectors, an iceWerx module with a Lattice iCE40HX8K
FPGA, and a Feather-compatible socket, which can take any number of
microcontroller boards such as the SparkFun RISC-V RED-V Thing Plus or the
Adafruit ESP32 Feather.  This board is designed to work with the Pumpkin
Space CubeSat Kit Frame, for example parts:
710-00794       Base Plate Assembly
703-00289       Chassis Walls
710-00713       Large Aperture Cover Plate Assembly
from https://www.pumpkinspace.com/